Added Testcases for IDE component
